+++ CHANGELOG.md
@@ -8,9 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Updated
-
 - Updated cards to use Next Image (DR-3056)
 - Removed canonical tag to old DC (DR-3264)
+- Refactored Adobe Analytics page names (DR-3257)
 
 ## [0.1.16] 2024-10-31